Move account code to it's own file
